zodiacSigns[] because answers class is gone
we were originally calling the zodiac list in the answers class to get the sign but since we're not using that anymore i put it in here. 
I also changed calculateUsersZodiacSign() to a public method that returns an int. please review the method getUsersZodiacSign() for duplicate code because of it.